Fixes Trie class tess; upgrades deps; refactors
