+++ src/filter/termConfig.ts
@@ -1,4 +1,7 @@
-// This will not scale well with localization and the size the term bank will grow
+/**
+ * This will not scale well with localization and the size the term bank will grow.
+ * This is a temporary solution until a better system can be thought out.
+ *  */ 
 export const termFilter = [
     "blockchain",
     "hodl",
