+++ src/filter/trie.ts
@@ -114,24 +114,4 @@ class Trie {
     }
 }
 
-
-class AutocompleteFilter {
-
-    trie: Trie;
-
-    constructor() {
-        this.trie = new Trie();
-    }
-
-    getFilter = (filterItems : string[]) => {
-
-        filterItems.forEach((item: string) => {
-            this.trie.insert(item.toLowerCase());
-        });
-
-        return this.trie;
-    }
-}
-
-
-export { AutocompleteFilter, Trie }
+export default Trie;
